fix: remove redundant u32 cast flagged by CI clippy (rust 1.93) (#174)
Local toolchains newer than the CI image's pinned rust:1.93 don't fire
clippy::unnecessary_cast on this pattern, which is how it slipped onto
main. Add a CLAUDE.md note to run clippy via the Docker CI container
before opening a PR so toolchain drift can't hide lint failures again.

+++ crates/oasis-backend-sdl/src/core_impl.rs
@@ -135,7 +135,7 @@ impl SdiCore for SdlBackend {
                     if ttf.has_glyph(ch) {
                         ttf.advance(ch, px).max(0) as u32
                     } else {
-                        oasis_types::bitmap_font::glyph_advance_scaled(ch, font_size) as u32
+                        oasis_types::bitmap_font::glyph_advance_scaled(ch, font_size)
                     }
                 })
                 .sum();
